+++ crates/gpui2/src/style.rs
@@ -110,7 +110,7 @@ pub struct Style {
     /// The mouse cursor style shown when the mouse pointer is over an element.
     pub mouse_cursor: Option<CursorStyle>,
 
-    pub z_index: Option<u32>,
+    pub z_index: Option<u8>,
 
     #[cfg(debug_assertions)]
     pub debug: bool,
